arch-design/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.28.0
pyyaml>=6.0
pydantic>=2.0.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
python-dotenv>=1.0.0

# MCP (Model Context Protocol) support (uncomment to enable MCP integration)
mcp>=1.0.0

# Diagram generation dependencies
# Note: diagrams-as-code CLI is broken (v0.0.4), using Python diagrams library directly
diagrams>=0.20.0
graphviz>=0.20.0 